test: namespace options to PodBuilder (#4534)
Previously, all resources produced by PodBuilder had `default` as
the K8s namespace.

Now, it will examine the manifest YAML and try to find an explicitly
referenced namespace. If it finds more than one, it'll error out
(this is a safeguard - nothing trips it today; if ever hit, we can
add an _explicit_ namespace override to use since there is no sensible
default in those scenarios).

If it doesn't find a single explicit namespace, it'll fallback to a
"context" namespace value if set. This allows simulating the case
where you've got a non-default namespace set in `kubeconfig`. If
that is also not set, it'll fallback to `k8s.DefaultNamespace` (aka
`default`).

The "cluster namespace" has to always be passed to PodBuilder if
you want to use it because there's no practical way for PodBuilder
to infer that from the `Manifest`. This is fine because there's
really only a few isolated places where we care about this, so
those tests can use it, while all other tests get generally more
consistent data for free.