engine: store Pod update times on RuntimeState
This was migrated to KubernetesDiscovery but both does not make
sense there as well as doesn't actually work because the build
controller needs to set the value (in engine state), which will
then just get overwritten by the API value (always zero).

The logic here matches the current behavior, which is only used
for crashed container detection and is a bit odd, but this is
just to address the current regression.