engine: store BaselineRestarts on RuntimeState
This is very similar to #4567 in terms of motivation. There's a
bit more changed here to clean up some unnecessary indirection in
the reconciler now that it's simpler.

The key part is to allow the build controller to adjust the baseline
restart for live updates, so that any restarts before the live
update get ignored. As a result, while it'd be nice to have the
KubernetesDiscovery API handle this, it needs to live in RuntimeState
(at least for now).