local: do not run serve_cmd on auto_init=False
If a `local_resource` had no `update_cmd` (i.e. just a `serve_cmd`)
but was set to `auto_init=False`, it would end up starting anyway
because the server controller saw that the dependency status was
`not_applicable` (technically true).

To fix this, the dependency status is set to `none` for local
resources with `auto_init=False` that have not yet had a build
(either from being manually triggered or having `TRIGGER_MODE_AUTO`
and a file change). Once a build has happened, it will resume
returning `not_applicable` to indicate that there is no longer a
reason to block on a dependency.

Fixes #4548.