docker: change how we model "image builds show up in the cluster immediately"
We used to treat this as a property of the cluster type + the container runtime.

But this made it impossible to support clusters that sometimes use your docker
runtime, and sometimes do not.

For examples, see:
- https://github.com/tilt-dev/tilt/issues/4587
- https://github.com/tilt-dev/tilt/issues/3654
- https://github.com/tilt-dev/tilt/issues/1729
- https://github.com/tilt-dev/tilt/issues/4544

This changes the data model so that "image builds show up in the cluster"
is a property of the Docker client, not of the cluster.

This should be much more flexible and correct, and help us support multiple
clusters.

Fixes https://github.com/tilt-dev/tilt/issues/4544