docker: support DOCKER_HOST=ssh://... remotes
Docker 18.09+ added support for SSH remote Docker connections.
The code to handle this isn't actually in the Moby client portions
that we use, but in the Docker CLI (which we already have a
dependency on).

This adapts some of the code from the Docker CLI that attempts
to create a `ConnectionHelper` if the host value is set. Currently,
the logic only creates a helper for `ssh://` hosts and returns nil
otherwise, in which case the former logic is used (i.e. pass the
host as-is to Moby). If a connection helper is created, it's used
to create a custom HTTP client/transport that will use SSH under
the hood.

Fixes #4603.