klog: suppress empty group-version errors
These error messages are benign and represent an uncommon yet
valid scenario where a group-version has no resource types.

This seems to be most common with the `prometheus-adapter` if
there are no external/custom metrics registered, at which point
the `external.metrics.k8s.io/v1beta1` (or `custom.`) group(s)
are empty.

The error/log does not impact Tilt operation but is confusing,
particularly during a `tilt down` where it's interleaved with
the actual Tilt output and makes it seem as though the operation
has failed.

Fixes #4610.