k8s: update PortForwardStatus after initializing forwards
There are 3 cases where status gets updated for a forward:
 * Forward started successfully (StartedAt set as well as local
   port, which is useful if a randomized one was requested)
 * Forwarder instance can't be created (bad config)
 * Forwarder returns an error after start

To simplify error handling and keep the status update logic
as centralized as possible, the actual initialization changed
slightly to happen in a single method.

There is some very basic logic to avoid spamming the API
server with updates on errors: it will only attempt to do
an update once a second if continuously failing. (Even then,
it will attempt to avoid a no-op update if the error hasn't
actually changed.)