api: add proxy for web UI to access apiserver
Use `kubectl proxy` code to provide the HUD with access to
apiserver routes.

Currently, this is restriction to JUST `UIButton` routes via
a path matcher on the handler. In the future, we'll use the
real apiserver authorization system here to configure an
unauthenticated/anonymous role for the UI for safe API access.