api: fill out PortForwardStatus type
It's possible to define multiple forwards on a single `PortForward`
spec, so the status conveys each of these independently, as it's
not possible to transactionally set up all forwards at once, meaning
some might be in a failure state while others are running.

It's also possible to let Tilt pick a random free port for the
local port, in which case, the status is also the only way to find
out what port got picked, so having per-forward granularity is
also necessary.