api: PodLogStream lifecycle from KubernetesDiscovery reconciler
Create + delete `PodLogStream` objects from `KubernetesDiscovery`
object status during reconcile based on discovered pods. This
eliminates the `runtimelog.PodLogManager` which was a store
subscriber that performed the same task - in fact, the data on
the store it was using was populated _by_ the `KubernetesDiscovery`
reconciler, so this brings it closer to the source and eliminates
a potential source of desync.