api: refactor reconciler initialization
Big change here is to split creation of the builder from the
actual starting of the reconciler so that all builders are
created FIRST and then once they're all configured, they get
started.

This is necessary because if a controller indexes a field,
it will throw an error because the shared informer has already
been started (by a previous controller) and you cannot add an
index once started.

After this change, all controllers can set up their field
indexes and any other changes safely.

At the same time, the `SetClient` pattern was removed - it
was no longer necessary because we introduced the `DeferredClient`
and the reconcilers were often actually getting that injected
anyway. Some light changes for the fixture to try to make things
more ergonomic accompany it, but it's still pretty awkward.