test: refactor PodLogStream reconciler tests
Eliminate the store part of the tests (besides using it to extract
logs).

This makes these tests purely about actual reconciler behavior and
not about implicit object creation from manifests; this will
temporarily reduce the test coverage, but will be reinstated by
tests for the `kubernetesdiscovery.Reconciler` to ensure that it
creates `PodLogStream` objects (instead of `runtimelog.PodLogManager`
which will go away). This is just split out because the scope of
changes required for these tests was sufficient on its own.