test: share K8s native PodBuilder
Several of the controllers interact with K8s Pods; the existing
PodBuilder used in manifest tests isn't appropriate as it's meant
to create Pods for manifests.

The PodLogStream reconciler has a good implementation of code to
generate realistic looking K8s pods, so moving it to the controller
`fake` package for use across reconcilers.