api: defer controller manager init
We've had a race condition here all along, but apparently the
latest changes exacerabated it (maybe just due to sheer number
of reconcilers we have now).

This moves the actual start of the controller manager to after
the controllers are initialized (which itself happens in two
phases).

It'd be nice to fully redo this initialization sequence in the
future, but we'd need to de-couple it from using the store to
initialize everything.