apis: add PodLogStreamTemplateSpec, PortForwardTemplateSpec
The basic types here should be non-controversial, and are something
you would want to have as a field on different types of objects.

The more controversial part of this PR is whether they should be fields of
KubernetesDiscovery, or of some more top-level type. I go back and forth on
this, but think we might end up playing around with different ownership models.