tiltfile: dc_resource and k8s_resource should both be additive
Before this change,
- If you had 2 k8s_resource calls for the same resource, it would be an error.
- If you had 2 dc_resource calls for the same resource, some properties
  would merge and some would overwrite.

This behavior was confusing.

This changes both functions to "aggregate" options.

Fixes https://github.com/tilt-dev/tilt/issues/4736
Fixes https://github.com/tilt-dev/tilt/issues/3738

I think that has much better semantics in a world with extensions that want to
"add" things to a resource, and is a good stopgap as we add more things
to the API server.