tiltfile: pass TILT_PORT to local() commands
Various Tilt extensions have been using `local()` to shell back
out to Tilt to use the API (e.g. `tilt apply -f ...`). This does
not play nice with custom ports if Tilt was invoked using the
`--port` argument.

Now, the current Tilt instance's port is passed as the `TILT_PORT`
argument so that the invoked command will automatically connect to
the same instance of Tilt.