tiltfile: pass TILT_HOST + TILT_PORT to local() commands (#4773)
Various Tilt extensions have been using `local()` to shell back
out to Tilt to use the API (e.g. `tilt apply -f ...`). This does
not play nice with custom ports if Tilt was invoked using the
`--port` argument.

Now, the current Tilt instance's port is passed as the `TILT_PORT`
argument so that the invoked command will automatically connect to
the same instance of Tilt.

Additionally, `TILT_HOST` is passed, which isn't needed for the
newer API commands, but IS needed for `tilt dump engine` and
`tilt logs`, for example.