probe: log transitions at higher level
This does not change any of the behavior/logic for probes.

Previously, ALL probe activity was logged at `Verbose`. Now,
transitions are logged at `Warn` (if transitioning TO failure)
or `Info` (if transitioning TO success). Sequential failures
continue to be logged at `Verbose` while sequential successes
continue to be suppressed.