docker-compose: use compose-go to parse Compose YAML (#4777)
Docker has formalized the Compose spec and has the `compose-go`
library to parse it. This is what the new v2 Compose CLI uses
itself (being activated on Docker for Mac as of the latest
releases).

Schema-wise, it is capable of parsing both v2 and v3 versions
of the spec as used by the `docker-compose` v1 CLI. That's all
we need it for!

The only notable behavior change is that the service iteration
order, while still respecting dependencies, is NOT influenced
by order within the YAML. This same change impacts the Compose
v2 CLI; in fact, before these changes, Tilt would be broken with
it as `docker-compose config --services` (what it was using)
is totally random on Compose v2! So, while this is a behavior
change, relying on YAML ordering was never a guarantee of Compose.

There was also a bug in Tilt's manual parsing with port mappings
that used a randomized host port, e.g. `3000/tcp`, where it
treated that as `3000:3000/tcp`, which is NOT the case. The only
impact of this was that previously we'd erroneously have shown a
non-functional endpoint link in the UI, which will now no longer
show.

As a bonus, this should speed up parsing operations that involve
Docker Compose, as we now do everything from native Go code instead
of shelling out 2x to Python!

But mainly, this will ensure that Tilt consistently parses the
Compose YAML (regardless of Docker Compose version), which can avoid
bugs like the above but also means unit tests won't start randomly
failing when Compose slightly changes the YAML output format (and
also leaving us in a weird situation where either every dev + CI
needs to upgrade or remain on an old version).

Tilt still shells out to the Compose CLI for other operations such
as streaming logs, as these are not part of the `compose-go` "spec"
library, and the `compose-cli` project is realistically not a
dependency we want :)