docker-compose: fix stack overflow panic (#4799)
Some of the `compose-go` types are not marshalable with `yaml.v3`:
they will trigger infinite recursion and a stack overflow panic.

Internally, `compose-cli` uses `yaml.v2` when marshaling (e.g. for
output via `docker-compose config`).

This is now explicitly used, and a test has been added to prevent
regression. Due to other issues/requirements, `compose-go` is also
in the midst of upgrading to `yaml.v3`, so in the future we should
be able to upgrade it and switch to `yaml.v3` for marshaling its
types.

Fixes #4797.