tiltfile: add exit() built-in to stop execution
Currently we have `fatal()` for terminating Tiltfile execution
with an error.

However, there are some circumstances where it's desirable to
stop Tiltfile execution but NOT trigger an error. For example,
which debugging a problem, it's useful to disable everything
subsequent in the Tiltfile.

This adds `exit()` which will stop execution in a "successful"
manner.

An optional message can be provided which will be printed before
exiting.

Closes #4501.