tiltfile: limit scenarios where `[no output]` is shown
Currently, `[no output]` is ALWAYS logged if a command from
`local()` / `helm()` / `kustomize()` returns no output.

This can be annoying with `local()` if you have a silent command
using `echo_off=True` + `quiet=True` because it'll result in
`[no output]` in the logs with no context.

At the least, this should be tied to `quiet` - if you've already
requested to NOT receive output, it's weird to get output when
there is no output.

This change also goes one step further - if `echo_off=True` and
`quiet=False`, it _also_ omits the `[no output]` because there's
no way to know what ran and didn't output anything since the
command itself wasn't echoed.