tiltfile: add exit() built-in to stop execution (#4819)
Currently we have `fatal()` for terminating Tiltfile execution
with an error.

However, there are some circumstances where it's desirable to
stop Tiltfile execution but NOT trigger an error. For example,
which debugging a problem, it's useful to disable everything
subsequent in the Tiltfile.

This adds `exit()` which will stop execution in a "successful"
manner.

An optional message can be provided which will be printed before
exiting. N.B. Unlike in Python, `exit()` is always a graceful/successful
exit, so `exit(1)` will log `1` but then exit without an error. In Tilt,
`fatal()` should be used to exit with errors.

A loaded/included Tiltfile can also call `exit()` which will halt
execution _entirely_ (i.e. of the root Tiltfile as well).

Closes #4501.