extensions: fix extension loading in non-git Tilt projects
Update to `tilt-dev/go-get@v0.2.1` which includes a small fix for
CWD when executing commands.

In general, this would result in a spurious git submodule update in
the user's Tilt project repo (i.e. wherever the `Tiltfile` lives)
but things would still work since the `tilt-extensions` repo does
not rely on submodules.

However, if the directory with the `Tiltfile` happened to not be a
git repo, the extension load would fail with a cryptic error about
it not being in a git repo:
```
fatal: not a git repository (or any of the parent directories): .git
```
This is because it was trying to do the submodule update
in the wrong place.