engine: don't look in completed pods for running containers
I was playing around with live-update Jobs.

I saw weird behavior where:
- builds would get stuck pending forever on already completed jobs
- tilt would think that a pod had crash when a job completed normally

Note that this changes the behavior of crash rebuild.
Before, if a pod crashed, Tilt would immediately begin a new build.
Now, if a pod crashes, Tilt waits for the controller to replace
it before starting the crash rebuild.
I think this behavior is probably more correct, given
that Tilt can't know if the controller is going to replace the Pod.