docker: exit early under more conditions in pruner
Recently received a pprof from a user who was experiencing high CPU
while Tilt wasn't really doing anything.

One of the offenders is the Docker Pruner: due to being an older
subscriber, it didn't have the newer optimization to skip running on
log-only store events, so would fire at a high-rate even when it
had no work to do.

Beyond that, collecting image references is surprisingly expensive
due to using a bunch of regex stuff under the hood. While it should
get called orders of magnitude less due to skipping log-only updates,
further changes are included to try and optimize this to exit early
without doing this work if possible.