docker: platform argument to customize image arch (#4934)
Docker CLI has a `--platform` argument and `DOCKER_DEFAULT_PLATFORM`
environment variable for default.

This is useful when dealing with multi-architecture images in
different scenarios, e.g. using an M1-based macOS machine.

This adds a `platform` argument to `docker_build` which will get
passed through to Docker/BuildKit. Similar to the Docker CLI, if
not specified, `DOCKER_DEFAULT_PLATFORM` will be used as a default
if set.

Fixes #4932.