docker-compose: log tailing robustness fixes (#4943)
The core issue this aims to solve is that for a stopped container,
`docker-compose logs` exits, and Tilt would continuously attempt
to re-launch it, resulting in high CPU usage due to repeatedly
invoking the command.

However, there were actually several intermingled issues that this
exposed/worsened. These mostly tie back to the behavior of
`docker-compose logs`: there is no built-in support for time
filtering (unlike `docker logs --since ...`), and since it exits
when a container stops, once invoked again, all the old logs will
be returned again. This means if a container crashes/restarts,
we'd re-emit all the old logs each time. Similarly, if there is
a "job" container that naturally runs to completion, whenever
re-triggered from the UI to run it, all the logs from prior
invocations would be re-logged. (This could also mean when attaching
on start up that you'd get a bunch of logs prior to the last
[re]start, which isn't super useful either.)

To solve this issue, `docker-compose logs` is now invoked with the
`--timestamps` argument, which prepends an RFC3339Nano timestamp
before each log line. Of course, there are caveats: first, Compose
v2 includes a whitespace before the timestamp and second, meta logs
from Compose itself about container events do NOT have a timestamp.
As a result, the attempt to parse/filter by the timestamp is _very_
forgiving. (There's an additional wrinkle which is that evidently
the `StartedAt` time reported by Docker for a container is not
precise, so you can actually get logs with earlier timestamps!)

Tying back to the original problem that once `docker-compose logs`
exits, Tilt tries to continuously re-invoke it, the watcher behavior
has been altered to more closely mirror the Pod log stream watcher.
When reading, if EOF is reached and/or the context is canceled,
it stops. The former case means that `docker-compose logs` exited
normally, indicating that the container is no longer running. The
latter case means the manifest is being torn down, so we no longer
want logs. Any errors during reading will cause it to retry, e.g.
if `docker-compose logs` crashes or is terminated abnormally.
(To support propagating process execution errors, the way that
logs are read has been adjusted slightly, which actually resolves
a race condition on reading logs and the `docker-compose logs`
process terminating, which ensures that short-lived containers
get all their logs!)

In the subscriber, when looking at already created watchers, if
the watcher has stopped running, a new one will only be created
if the corresponding container's start time is AFTER the last
time it started. This is the key to not continuously starting
up new `docker-compose logs` processes! (Finally.) We can't
key off container ID because that's consistent across restarts,
including both crashes and job containers that are run multiple
times without changes.

Finally, there is a fix for the Docker Compose event watcher
which was inadvertently broken (by yours truly) a while back.
It was not actually _ever_ running, so Tilt wouldn't notice
external changes (e.g. `docker-compose restart foo`), which
meant state was out of sync. This was not really noticeable
before since we were insanely aggressive about re-establishing
`docker-compose logs` sessions. After these fixes, however,
it meant that any container events _not_ initiated via Tilt
would go unnoticed, so the logs would not get resumed, as the
new container start time would never get picked up.

TL;DR Docker Compose logs should behave much nicer now both
      in terms of CPU usage as well as under various common
      scenarios like container restarts or job containers