engine: handle all link types + don't invalidate build on change
In some cases, `url.URL` can't be compared: case reported was a
URL with a `user@` portion, which populates the `User` field of
type `url.Userinfo` that contains unexported fields. (This came
up when trying to add a link for a local DB connection string.)

While one possible fix would be to add a custom comparer for
`Link` which uses `Link::URLString()` to side-step this, since
the comparers are used for build invalidation, we can actually
just ignore the `Link` type entirely for comparison purposes,
as it does not cause a build to be invalidated.

So, as a bonus, you can now freely change `links` in `Tiltfile`
without triggering new builds unnecessarily!

Fixes #4945.