engine: fix flaky Docker Compose test (#4955)
This is not the result of any behavior changes from the recent
Docker Compose changes. This test is a negative test case to
ensure we filter out the "Attaching to foo" message Docker Compose
emits at the beginning of every log session. Previously, it was
only checking that it hadn't seen that exact message. In attempt
to tighten it up, it now asserts that it hasn't seen that message
AND that it HAS seen a message that _should_ be output.

Turns out that there was a race in the test, so it was looking at
the logs while they were still empty sometimes. It would pass
before but fails now due to being stricter.

The solution is a copy-paste from another DC test.