cli: add Docker Compose version to `tilt doctor` output (#4963)
This uses the version detection/parsing for Docker Compose from
#4962 to include the Compose version in `tilt doctor` output. (This
will be particularly useful for pinpointing v2 compatibility
issues.)

Due to the nature of how we get the version info, all errors are
ignored and the field will be omitted. In particular, it'll be
very common for the `Version()` method to return an error because
the user doesn't even have Compose installed, so it has nothing to
shell out to, and the errors returned from command failures are
not friendly towards the `tilt doctor` output.