telemetry: upgrade to otel 1.0 (#4978)
There's a substantial number of changes here due to moving from
v0.2.0 to v1.0.0. Notably, the way that we were marshaling spans
to JSON doesn't work anymore, so there's a decent amount of bridge
code, which is a fork of the structs and marshaling code as of
v0.2.0 with helper functions to convert from the newer types into
them.

Eventually, we should probably provide an option to the
`experimental_telemetry_cmd` (or replacement) to control the
serialization format and use the canonical OTLP JSON format (but
this is not yet finalized).

Beyond this, the span collector is now a `SpanExporter` rather
than a `SpanProcessor` and is invoked by the built-in batch
processor. This actually impacts tests more than production code!