demo: add k3d-via-Docker client to provision clusters
To simplify trying out Tilt when following the new tutorial, we
want an "all-in-one" `tilt demo` command that will launch a
temporary K8s cluster, clone the sample project, and `tilt up`
against it.

To avoid extra dependencies within Tilt, this is an experimental
k3d client that executes k3d commands by running k3d itself via
Docker with the Docker socket mounted. This eliminates the need
for the user to have k3d downloaded (or Tilt to play "package
manager" and grab the appropriate binary from GitHub releases.)

There's a few quirks/limitations with this approach, namely that
k3d can't interact with the host system to e.g. update kubeconfig.
However, that's actually desirable in this instance as we want to
keep things isolated.

Longer-term, particularly if we want Tilt to truly manage clusters
as part of the dev environment (that are meant to be more robust
and "permanent" than for a tutorial), it'd probably make sense to
use `ctlptl` as a library and explore options within `ctlptl` to
avoid requiring `k3d` itself as a dependency (note: `ctlptl` does
not actually support `k3d` today).