script: update Toast configuration for Go 1.16+ (#5009)
* script: update Toast configuration for Go 1.16+

We've been using an old image for Toast tasks based on Go 1.12
(the Dockerfile was updated but the digest still pointed to the
old one).

Things broke recently because `toast wire` was touching code that
required Go 1.16+ so couldn't compile anymore.

There's a bunch of annoying changes required for the protobuf stuff
to work, but everything is in order now. The versions of packages
here are newer than before, so the tasks have been re-run. The
generated protobuf code looks almost entirely different, for example.