controllers: use FileWatch to filter for ignores
This is a step towards FileWatch directly driving LiveUpdate
behavior (i.e. actually using it to determine changed files).

For now, it's just used to generate the ignore list, which
keeps things more in sync vs. generating a separate filter.

The main benefit to this is actually the test changes, which
ensure the FileWatch object for the image now consistently
exists, so they'll be better prepared for additional reliance
on it in the future.