docker: pass the correct path to Tiltfile
This was using the workdir rather than the Tiltfile path. The
only thing it's used for is setting up an ignore for the Tiltfile.

In practice, this should have meant that image builds would be
totally broken - it'd ignore the _entire_ build context directory.
However, there's an extra bit of logic when actually tar'ing things
for the build context that explicitly prevents you from ignoring
the root directory :)

This will now prevent `Tiltfile`s from being included in build
contexts, and shortly the corresponding `FileWatch` for an image
will ignore changes to the `Tiltfile` so that it doesn't spuriously
trigger.