docker: evaluate build args for base image ref (#5048)
This was 90% wired up and already worked in the case that the
`ARG` referenced in `FROM` had a (relevant) default value.

However, we weren't passing the _actual_ build args values that
come from `build_args` arg in `docker_build` (akin to `--build-arg`
in Docker CLI).

For ergonomics and to avoid leaking Docker internals, these are
passed in as a `map[string]string`, which mirrors Tilt's internal
usage. Order does not matter for these given the map guarantees
uniqueness of arg name (key), so there's no precedence logic here.
(However, once actually used, the values from the AST are prepended
so that they have lower precedence than Tiltfile values.)

Fixes #5045.