k8s: more robust CRD replacement
Currently, there are CRDs that are so long that they blow up kubectl apply.
https://github.com/strimzi/strimzi-kafka-operator/issues/4589

Delete+Create is a bad strategy for handling this problem.
Deleting the CRD deletes all the resources that depend on it,
which is not what we want.

Instead, add a 3rd strategy for upserting a resource that
allows existing custom resources to persist.