controllers: eliminate LU filter + export path mapping logic (#5059)
Per discussion on #5026, there's no need for an additional filter
when tar'ing the files for LiveUpdate as they've already been
filtered, so this was purely duplicative.

The other change here is just to slightly refactor the logic
to evaluate a set of changed files against a `LiveUpdateSpec`
and export it so that it can be used by the reconciler (against
`FileWatchStatus` directly vs implicitly in engine). Those
changes are forthcoming independently, this is just a quick
bit to keep PRs small.