api: add stub handler for KubernetesApply Cmd deploy
This is mostly just moving a few things around to add a stub for
performing a deploy using a `KubernetesApplyCmd` instead of direct
YAML.

Currently, if a `KubernetesApply` object has a `Cmd` in place of
`YAML`, it will always fail with an error about not being supported
yet.

There is a behavior change to support the forthcoming functionality,
however, which is that on deploy/apply failure, GC of old objects
won't occur. This was okay for YAML because we had the fully composed
version of what we were _trying_ to apply ahead of time, so even if
it failed, we could still try to delete things that are no longer
referenced. In the command case, it does an apply and then returns
the applied YAML on success, so if it fails, there's no way to know
what it attempted to apply. As a result, the easiest thing is to
not attempt any GC if deploy mechanism fails.