api: add RestartOn support to KubernetesApply reconciler (#5094)
Currently, this code path won't be hit since Tiltfile-managed
KubernetesApply objects won't have a `RestartOnSpec`.

However, this will be important for the forthcoming logic to
support a custom `Cmd` to perform the apply + generate YAML,
so that the `Cmd` can be re-run when its inputs change. In the
future, this will also make it possible to wire up a "force
re-deploy" action in the UI (i.e. the trigger resource button,
but API-native).