localexec: provide a facade for local command exec
This is a Tilt-centric local process executor meant for commands
that run to completion. The `Cmd` controller, which manages
long-running processes for `local_resource`'s `serve_cmd`, has
its own unexported implementation that's focused more on async
signaling of process status for a daemon-style process that runs
perpetually.

There is overlap between these two, specifically in that they
take care to use process groups and terminate the entire group to
avoid leaking resources.

The Tiltfile loader has been updated to use this for running
`local()` commands, which should be invisible to consumers, but
does mean that it's now possible to inject a fake execer for
Tiltfile loading and return pre-canned responses for `local()`
calls, which could be useful.