localexec: improve flaky test (#5115)
This failed a couple times in CI - I can't reproduce the failure
on my macOS laptop, but I'm guessing it's a timing issue due to
the CI runners being slower.

Changes are two-fold:
 * Aggressively cancel context (i.e. trigger SIGKILL to process
   group) as soon as PID is seen - this makes it possible to run
   this test hundreds of times much quicker, which was useful to
   try and reproduce the issue
 * Allow grace period for child PID to get terminated - hopefully
   this prevents flakiness on CI runners

It's _possible_ the CI flakiness is actually due to the child PID
getting recycled, but that seems very unlikely given the timings
here combined with standard OS behavior for Linux. If these changes
don't eliminate the flakiness, we might just need to reduce the
strictness of this test and not attempt to assert on the child PID
at all (see `cmd/excer_test.go`, which has a similar test and does
not attempt to assert on the child process state).