api: simplify/unify RestartOn initialization
Replace `ExtractKeysForIndexer` + `RegisterWatches` with a unified
`restarton.SetupController` function, which uses these (now unexported)
functions internally.

This is expected to be called as part of a reconciler's `CreateBuilder()`
method and will add watches for all restart on types as well as set up
the indexer. A func must be passed to pull the `RestartOnSpec` and/or
`StartOnSpec` objects off the reconciler object type.

The `ControllerFixture` now invokes `CreateBuilder` to ensure this logic
gets called during tests, which sometimes rely on the indexer being
properly set up. Note that no automatic reconciliation for dependent
objects will happen in a test scenario still.

The `Cmd` reconciler tests were refactored to use `ControllerFixture`
(they pre-dated its existence), so that they continue to pass given
the new functionality describe above which they require.