api: fix race during FileWatch cancellation
It's possible for the monitor to get cancelled while in the midst
of reporting an event, which will cause the API operations to fail.

We don't want to treat these as fatal errors: since the monitor
was cancelled, the failed update was stale anyway and we should
just silently ignore it.