api: fix race during FileWatch cancellation (#5139)
It's possible for the monitor to get cancelled while in the midst
of reporting an event, which will cause the API operations to fail.

We don't want to treat these as fatal errors: since the monitor
was cancelled, the failed update was stale anyway and we should
just silently ignore it.

NOTE: There's a similar change made for the `Cmd` reconciler,
but it would only have been triggered on shutdown, so wouldn't
have had any visible impact. However, the stricter fake API client
exposed it in tests, causing them to fail non-deterministically, so
it's changed here as well.