tiltfile: add cmd() built-in
Tiltfile functions that accept commands currently require 4 arguments
per command (cmd, env, dir, cmd_bat). In some cases, like
`local_resource` where there are multiple commands, that gets unwieldy
fast.

The `cmd` built-in lets us group these into a single object, which
maps to the `model.Cmd`. Anywhere that currently accepts a `cmd`
can now optionally accept this instead. However, to avoid breaking
Tiltfile compatibility, the separate arguments are still available,
but CANNOT be used in conjunction with the `cmd()` built-in for a
given command assembly.

There's a side benefit here when working with API types in a Tiltfile,
which is that this can be used as a bridge/helper to get things into
the expected format. For example, the APIs all take `Args` as an
argv-style string array, and don't support the auto-magical
string -> shell wrapper, so callers can now make a `cmd` object and
then use `cmd.args` for the field value. Ditto for env, where it's
nicer in Starlark to make a `dict`, but that then needs to be mapped
to a `KEY=value` string list, which can be accessed via `cmd.env`.