api: run DeleteCmd when KubernetesApply object deleted
When a custom `DeployCmd` is provided, a corresponding `DeleteCmd`
is needed to handle cleanup.

Although Tilt is able to track + diff the managed entities to determine
what to delete (which is still the case when using YAML), the
`DeployCmd` might create additional state, e.g. a Helm release in
Tiller/`Secret`, so this provides the tool with the means to clean
up the external state _as well as all relevant entities in the
cluster_.