tiltfile: add delete_cmd to k8s_custom_deploy
This wires up the Tiltfile + `tilt down` side of things (the API
support is in #5152).

The `delete_cmd` is required by `k8s_custom_deploy` since otherwise
there's no way to clean up resources via `tilt down`.

The feature flag is also removed _entirely_ - it was added as a
guard pending this change, but an interim Tilt release never
actually shipped, so there's no need to obsolete it first.